command: Better visual hierarchy for diagnostics
I frequently see people attempting to ask questions about Terraform's
error and warning messages but either only copying part of the message or
accidentally copying a surrounding paragraph that isn't part of the
message.

While I'm sure some of these are just "careless" mistakes, I've also
noticed that this has sometimes overlapped with someone asking a question
whose answer is written directly in the part of the message they didn't
include when copying, and so I have a theory that our current output
doesn't create a good enough visual hierarchy for sighted users to
understand where the diagnostic messages start and end when we show them
in close proximity to other content, or to other diagnostic messages.
As a result, some folks fail to notice the relevant message that might've
answered their question.

I tried a few different experiments for different approaches here, such
as adding more horizontal rules to the output and coloring the detail
text differently, but the approach that felt like the nicest compromise
to me was what's implemented here, which is to add a vertical line
along the left edge of each diagnostic message, colored to match with the
typical color we use for the diagnostic's severity. This means that the
diagnostics end up slightly indented from what's around them, and the
vertical line seems to help subtly signal how we intended the content
to be grouped together.

I originally tried this with Unicode box drawing characters, which gave
a nice result. However, we're currently still constrained by supporting
Windows 8's legacy Windows Console, which has no Unicode support and so
cannot handle these box drawing characters.

As a compromise for now I've switched them out for similar ASCII-art-style
characters, which is not ideal because it's ambiguous in non-visual
terminals and it's not as elegant looking (subjectively), but I think
overall this is a good tradeoff for the current constraints and that
non-visual terminals will encounter similar sorts of output from other
software and thus should have heuristics to handle it somewhat gracefully.
Hopefully it won't be long before we can officially end our Windows 8
support and thus require a modern terminal which supports Unicode
properly, and then we can swap back to the box drawing characters.